+++ resources/views/emails/user/approval.blade.php
@@ -1,0 +1,12 @@
+@component('mail::message')
+# Richiesta di approvazione
+
+Caro amministratore, l'utente {{ $user->name }} di {{ $user->horecaname }} chiede di essere eletto al ruolo di Publican.
+
+@component('mail::button', ['url' => $url])
+Utenti
+@endcomponent
+
+Grazie,<br>
+{{ config('app.name') }}
+@endcomponent